DanaR E2E: extract AbstractDanaEmulatorUiTest so DanaR can reuse the UI flow
Pulls the pump-agnostic UI driving out of DanaRsEmulatorUiTest into an abstract
base (find/click/openVia/openManageAction, the whole insulin-delivery-from-UI
flow, queue-idle logic, profile/app seeding), leaving DanaRsEmulatorUiTest a thin
subclass that supplies the RS-specific hooks (transport/pairing/active-pump
seeding, emulator-state reads, teardown, status-read) plus its BLE5-only screen
walk. Behaviour is unchanged - verified on emulator-5554: BLE5 full walk, v3 and
v1 delivery all pass through the refactored base (v1's known intermittent
deferred-callback flake aside; its disconnect-before-await teardown mitigation is
preserved).

Next: a DanaREmulatorUiTest subclass reusing the same flow over RFCOMM.
